Add PouchDB (using SQLite under the hood) as an alterantive to Firebase.
The issue with firebase is that it relies on IndexDB as the local DB,
but React Native doesn't support that. Another alternative here was to
use the reat-native-forebase library but that needs to eject from expo.

So, I brought in PouchDB + SQLite which is provided by expo, and works without
ejecting.